requirements.txt
fastapi[standard]
uvicorn[standard]
pydantic
sqlalchemy
alembic
python-dotenv
pytest
requests
librosa
deeprhythm
numpy